20250125update

+++ app/controllers/favorites_controller.rb
@@ -9,8 +9,10 @@ class FavoritesController < ApplicationController
 
   def destroy
     book = Book.find(params[:book_id])
-    favorite = current_user.favorites.new(book_id: book.id)
-    favorite.destroy
+    favorite = current_user.favorites.find_by(book_id: book.id)
+    if favorite
+     favorite.destroy
+    end
     redirect_to book_path(book)
   end
